Move spell checking handling to renderer process
